+++ src/main/kotlin/com/sparta/techTree/comment/service/CommentService.kt
@@ -4,11 +4,11 @@ import com.sparta.techTree.comment.dto.CommentDTO
 import com.sparta.techTree.comment.dto.UpdateCommentRequest
 
 interface CommentService {
-    fun updateComment(commentId: Long, userId:Long, request: UpdateCommentRequest): CommentDTO
+    fun updateComment(commentId: Long, userId: Long, request: UpdateCommentRequest): CommentDTO
 
-    fun deleteComment(commentId: Long, userId:Long)
+    fun deleteComment(commentId: Long, userId: Long)
 
-    fun getCommentsByPost(postId: Int): List<CommentDTO>
+    fun getCommentsByPost(postId: Long): List<CommentDTO>
 
     fun createComment(commentDto: CommentDTO): CommentDTO
